% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm_importaDFP.R
\name{cvm_importaDFP}
\alias{cvm_importaDFP}
\title{Importa dados da CVM}
\usage{
cvm_importaDFP(
  grupoconta = "ativo",
  tipo = "consolidado",
  metDFC = "direto",
  ano = 2020,
  path = NULL
)
}
\arguments{
\item{grupoconta}{Demonstrações contábeis: 'ativo', 'passivo' e 'dfc'}

\item{tipo}{'consolidado' ou 'individual'}

\item{metDFC}{'direto' ou 'indireto'}

\item{ano}{Ano da tabela que se deseja realizar a extração}

\item{path}{A pasta (link) em que se encontra o conjunto das tabelas}
}
\description{
Importa dados da CVM
}
\details{
A função possui a utilidade de realizar a extração de dados do repositório da CVM,
a partir de 5 parâmetros escolhidos pelo usuário.
}
\examples{

library(cntdd)

cvm_importaDFP()

# Da forma acima, será chamada automaticamente os parametros: ativo', 'consolidado', 'direto', '2020' e NULL

Outra forma de se chamar a função:

cmv_importaDFP(grupoconta = "passivo", tipo = "individual", "indireto","2019")


}
