% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtCnpjCodB3.R
\docType{data}
\name{dt_cnpjCodB3}
\alias{dt_cnpjCodB3}
\title{CNPJ vs Codigos da B3 e CVM}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1352 rows and 3 columns.
}
\usage{
dt_cnpjCodB3
}
\description{
Lista de CNPJs com códigos da CVM e da B3
}
\details{
Data.frame com 1.352 linhas e 3 colunas.
}
\keyword{datasets}
