% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtContabil.R
\docType{data}
\name{dt_contabil}
\alias{dt_contabil}
\title{Dados das empresas}
\format{
An object of class \code{data.frame} with 3 rows and 6 columns.
}
\usage{
dt_contabil
}
\description{
Dados contabeis das empresas
}
\keyword{datasets}
