% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcvmB3.R
\docType{data}
\name{dt_cvmB3}
\alias{dt_cvmB3}
\title{Dados cadastrais das empresas da CVM}
\format{
Data.frame com 1.229 observações (linhas) e 13 variáveis (colunas).
}
\usage{
dt_cvmB3
}
\description{
Lista de empresas cadastradas na CVM com seus respectivos CNPJs, código CVM e
código de negociação na B3. Inclui dados de constituição, setor, dentre outros.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{dt_cvmB3}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
