% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_iseB3.R
\docType{data}
\name{dt_iseB3}
\alias{dt_iseB3}
\title{Conjunto de Dados ISE B3 (2021-2023)}
\format{
Data.frame com 40 linhas e 35 colunas.
}
\usage{
dt_iseB3
}
\description{
Dados de desempenho ESG das empresas participantes dos processos de seleção
às carteiras 2021 e 2023 do ISE B3 disponibilizadas no módulo "ESG Índices
da Plataforma ESG Workspace.
}
\details{
Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub}.
Ao acessar, fazer busca pelo nome da funcao \code{dt_contabil}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
