% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmeses.R
\docType{data}
\name{dt_meses}
\alias{dt_meses}
\title{Meses}
\format{
An object of class \code{data.frame} with 12 rows and 1 columns.
}
\usage{
dt_meses
}
\description{
Dados de meses e suas referências em relação a bimestre, semestre
}
\keyword{datasets}
