% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_analisaDRE.R
\name{ind_analisaDRE}
\alias{ind_analisaDRE}
\title{DRE Detalhada}
\usage{
ind_analisaDRE(
  empresa = "Empresa",
  periodo = (year(Sys.Date()) - 1),
  receita = 150,
  custo = 30,
  despSemDep = 20,
  depreciacao = 10,
  juros = 35,
  ircsll = 15,
  dividendos = 10,
  aliquotaIR = 0.34,
  verDados = F,
  fontBar = 3,
  colorBar1 = "#9ecae1",
  colorBar2 = "#3182bd",
  colorBar3 = "#e34a33",
  colorBar4 = "#fee8c8",
  ajustLabel = 8
)
}
\arguments{
\item{empresa}{Um vetor tipo character com o nome da empresa}

\item{periodo}{Vetor numérico ou string indicando o período da análise}

\item{receita}{Vetor com os valores da receita líquida da empresa}

\item{custo}{Vetor com os valores do custo da empresa}

\item{despSemDep}{Vetor com os valores das despesas operacionais sem o valor da depreciação}

\item{depreciacao}{Vetor com os valores da depreciação}

\item{juros}{Vetor com os valores dos juros passivos}

\item{ircsll}{Vetor com os valores dos impostos incidentes sobre o lucro}

\item{dividendos}{Vetor com os valores dos dividendos propostos}

\item{aliquotaIR}{Alíquota do Imposto de Renda (Padrão: 0,34 ou 34\%)}

\item{verDados}{Mostra data.frame? (TRUE/FALSE). Se \code{TRUE}, Mostra o data.frame. Se \code{FALSE},
mostra o gráfico (Padrão: \code{F})}

\item{fontBar}{Valor para ajuste do tamanho da fonte do texto sobre as barras (Padrão: 4)}

\item{colorBar1}{String (texto) com a especificação da cor da barra do tipo 1}

\item{colorBar2}{String (texto) com a especificação da cor da barra do tipo 2}

\item{colorBar3}{String (texto) com a especificação da cor da barra do tipo 3}

\item{colorBar4}{String (texto) com a especificação da cor da barra do tipo 4}

\item{ajustLabel}{Valor para ajuste da localização do texto sobre as barras (Padrão: 10)}
}
\description{
Essa função ajuda na análise da Demonstração do Resultado do Exercício (DRE),
demonstrando os diversos tipos de lucro com sua respectiva formação.
}
\details{
O gráfico gerado nessa função dar suporte na compreensão dos conceitos de
Lucro Bruto, EBITDA, EBIT, NOPAT, Lucro antes dos impostos, lucro líquido e
lucros retidos.

Apresenta como resultado um gráfico ou um data.frame com as contas para análise.
Se o parâmetro \code{verDados} for \code{FALSE} ou \code{F}, mostra um gráfico com a
formação do lucro da empresa. Se for \code{TRUE} ou \code{T}, mostra o data.frame com
as contas que compõem o gráfico.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.com.br/}.
Ao acessar, fazer busca pelo nome da função \code{ind_analisaDRE}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_analisaDRE(
empresa = "Empresa",
periodo = 2023,
receita = 80,
custo = 20,
despSemDep = 20,
depreciacao = 5,
juros = 2,
ircsll = 1,
dividendos = 2,
aliquotaIR = 0.34,
verDados = FALSE,
fontBar = 4,
colorBar1 = "#9ecae1",
colorBar2 = "#3182bd",
colorBar3 = "#e34a33",
colorBar4 = "#fee8c8",
ajustLabel = 10
)

}
