% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_analiseCiclo.R
\name{ind_analiseCiclo}
\alias{ind_analiseCiclo}
\title{Ciclo Operacional e Financeiro}
\usage{
ind_analiseCiclo(
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  pmrv = c(60, 75),
  pmre = c(45, 20),
  pmpc = c(50, 100),
  periodoGrafico = year(Sys.Date()) - 1
)
}
\arguments{
\item{periodo}{Vetor numérico indicando o período da análise}

\item{pmrv}{Vetor com os valores do prazo médio do recebimento de vendas}

\item{pmre}{Vetor com os valores do prazo médio da renovação dos estoques}

\item{pmpc}{Vetor com os valores do prazo médio do pagamento de compras}

\item{periodoGrafico}{Vetor com um dos periodos informado no parâmetro \code{periodo} a ser considerado na análise}
}
\description{
Essa função dá suporte na análise dos ciclos operacional e financeiro das empresas.
}
\details{
Apresenta como resultado um gráfico de barras com a representçaão gráfica
do prazo médio de renovação dos estoques, prazo médio de recebimento de vendas,
prazo médio de pagamento de compras, ciclo operacional e ciclo financeiro.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{ind_dupont}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_analiseCiclo(
 periodo = 2022:2023,
 pmrv = c(60, 75),
 pmre = c(45, 20),
 pmpc = c(50, 100),
 periodoGrafico = 2023
 )

}
\seealso{
\code{\link[=ind_pmre]{ind_pmre()}}, \code{\link[=ind_pmrv]{ind_pmrv()}}, \code{\link[=ind_pmpc]{ind_pmpc()}}, \code{\link[=ind_cicloOperacional]{ind_cicloOperacional()}}, \code{\link[=ind_cicloFinanceiro]{ind_cicloFinanceiro()}}
}
