% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_analiseDupont.R
\name{ind_analiseDupont}
\alias{ind_analiseDupont}
\title{ROE desmembrado (Dupont)}
\usage{
ind_analiseDupont(
  intervalo = 2009:2011,
  patrimonioLiquido = c(200, 150, 175),
  ativoTotal = c(300, 500, 600),
  receita = c(500, 450, 400),
  lucroLiquido = c(8, 10, 14),
  apenasDados = F
)
}
\arguments{
\item{intervalo}{Vetor numérico indicando o período da análise}

\item{patrimonioLiquido}{Vetor com os valores do patrimônio líquido da empresa}

\item{ativoTotal}{Vetor com os valores do ativo total da empresa}

\item{receita}{Vetor com os valores da receita líquida da empresa}

\item{lucroLiquido}{Vetor com os valores do lucro líquido da empresa}

\item{apenasDados}{Se \code{TRUE}, Mostra apenas o banco de dados. Se \code{FALSE}, mostra
o banco de dados e o gráfico com a análise dupont. (TRUE/FALSE)}
}
\description{
Essa função demonstra o desmembramento do ROE de acordo com a análise Dupont.
}
\details{
Apresenta como resultado uma lista com 2 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{plot} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução do ROE da empresa e seus componentes, conforme análise Dupont, durante
os períodos. Se for \code{FALSE} ou \code{F}, o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador. o banco de dado está no formato tibble que pode ser
usados individualmente durante o processo de análise de dados.
}

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.com.br/}.
Ao acessar, fazer busca pelo nome da função \code{ind_dupont}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_analiseDupont(
 intervalo = 2009:2011,
 patrimonioLiquido = c(200, 150, 175),
 ativoTotal = c(300, 500, 600),
 receita = c(500, 450, 400),
 lucroLiquido = c(8, 10, 14),
 apenasDados = FALSE
 )

}
\seealso{
\code{\link[=ind_roe]{ind_roe()}}, \code{\link[=ind_roa]{ind_roa()}}, \code{\link[=ind_nivelEndividamento]{ind_nivelEndividamento()}}, \code{\link[=ind_giro]{ind_giro()}}, \code{\link[=ind_ros]{ind_ros()}}, \code{\link[=ind_pmre]{ind_pmre()}}
}
