% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_analiseVertical.R
\name{ind_analiseVertical}
\alias{ind_analiseVertical}
\title{Análise Vertical}
\usage{
ind_analiseVertical(
  contas = c("estoques", "imobilizado"),
  valores = c(300, 700),
  contaBase = "Ativo Total",
  vrContaBase = 2000
)
}
\arguments{
\item{contas}{Um vetor tipo character com o nome das contas para análise vertical}

\item{valores}{Vetor numérico indicando o valor das contas para análise vertical}

\item{contaBase}{String com o nome da conta que servirá de base para a análise vertical}

\item{vrContaBase}{Valor numérico da conta base}
}
\description{
Essa função calcula a análise vertical de contas específicas.
}
\details{
Apresenta como resultado um data.frame com 4 colunas: conta, valor, AV e Conta Base.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_analiseVertical}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_analiseVertical(
contas = c("estoques", "imobilizado"),
valores = c(300, 700),
contaBase = "Ativo Total",
vrContaBase = 2000
)

}
