% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_analiseVH.R
\name{ind_analiseVH}
\alias{ind_analiseVH}
\title{Análise Vertical e Horizontal}
\usage{
ind_analiseVH(
  .dt = dt_contabil,
  .contas = c("caixaEquiv", "clientesCP"),
  .contaReferenciaAV = ativoTotal,
  .colEmpresa = empresa,
  .colPeriodo = ano,
  .output = "AVAH"
)
}
\arguments{
\item{.dt}{data.frame com, pelo menos, 4 colunas para preenchimento dos demais argumentos desta função}

\item{.contas}{Vetor tipo character com o nome das colunas referente às contas de interesse na análise}

\item{.contaReferenciaAV}{Nome da coluna com a conta de referência para a análise vertical}

\item{.colEmpresa}{Valor Nome da coluna que identifica a empresa (pode haver mais de uma empresa no banco de dados)}

\item{.colPeriodo}{Valor Nome da coluna que identifica o período da análise (pelo menos 2, por empresa, para a análise horizontal)}

\item{.output}{Informar \code{full} para o banco de dados completo, \code{AV} para mostrar apenas as análises verticais, \code{AH} para
mostrar apenas as análises horizontais, \code{AHAV} para mostrar, respectivamente, a análise horizontal e vertical ou
\code{AVAH} para mostrar, respectivamente, a análise vertical e horizontal}
}
\description{
Essa função calcula a análise vertical e horizontal de contas específicas de
um data.frame.
}
\details{
O usuário deve dispor de uma base de dados com, no mínimo, 4 colunas (variáveis):
\enumerate{
\item Coluna com identificação da empresa.
\item Coluna com identificação do período da análise contendo, pelo menos 2 períodos,
para cálculo da análise horizontal.
\item Coluna da variável de interesse para a análise vertical
\item Coluna da variável de referência para a a análise vertical.
}

Apresenta, como resultado, um data.frame com todas as colunas informadas, com
a análise horizotal ou com a anaálise vertical.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.com.br/}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_analiseVH}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_analiseVH(
.dt = dt_contabil,
.contas = c("caixaEquiv", "clientesCP"),
.contaReferenciaAV = ativoTotal,
.colEmpresa = empresa,
.colPeriodo = ano,
.output = "AVAH"
)

}
