% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_capitalInvestido.R
\name{ind_capitalInvestido}
\alias{ind_capitalInvestido}
\title{Capital Investido}
\usage{
ind_capitalInvestido(
  indicador = "capitalInvestido",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  patLiq = c(100, 300),
  empFinCP = c(150, 200),
  empFinLP = c(230, 190),
  atvTotal = c(500, 400),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numerico indicando o periodo da analise}

\item{patLiq}{Vetor com os valores do patrimonio liquido da empresa}

\item{empFinCP}{Vetor com os valores de emprestimos e financiamentos de longo prazo da empresa}

\item{empFinLP}{Vetor com os valores de emprestimos e financiamentos de curto prazo da empresa}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra grafico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatorio do indicador. Se \code{FALSE}, mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa funcao calcula, de forma objetiva, o capital investido. Corresponde ao
investimento realizado por acionistas (socios) e credores (bancos).
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Grafico} se o parametro \code{plot} for \code{TRUE} ou \code{T}, mostra um grafico com a
evolucao do capital investido da empresa durante os periodos. Se for \code{FALSE} ou \code{F},
o grafico nao e apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
calculo do indicador;
\item \strong{Indice} o capital investido dos periodos informados;
\item \strong{Analise Vertical} Analise Vertical das contas informadas no item 1. Contas de resultado
terao sua analise vertical em relacao a receita total e contas patrimoniais terao
sua analise vertical em relacao ao ativo total;
\item \strong{Analise Horizontal} Analise Horizontal das contas informadas no item 1.
}

Todos os itens da lista sao bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de analise de dados.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_roa}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
