% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_capitalOperLiquido.R
\name{ind_capitalOperLiquido}
\alias{ind_capitalOperLiquido}
\title{Capital Operacional Líquido (COL)}
\usage{
ind_capitalOperLiquido(
  indicador = "Liq Seca",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  patLiq = c(150, 200),
  investimentos = c(5, 3),
  dividasCP = c(30, 40),
  dividasLP = c(60, 90),
  atvTotal = c(500, 400),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{patLiq}{Vetor com os valores do patrimônio líquido da empresa}

\item{investimentos}{Vetor com os valores dos investimentos da empresa}

\item{dividasCP}{Vetor com os valores de empréstimos e financiamentos de
curto prazo (Passivo Circulante)}

\item{dividasLP}{Vetor com os valores de empréstimos e financiamentos de
longo prazo (Passivo Não Circulante)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatório do indicador. Se \code{FALSE},
mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa função calcula o Capital Operacional Líquido de forma simplificada, pois
não calcula o ativo e passivo operacional. Faz uma simplificação pela
soma do patrimônio líquido com a dívida de curto e longo prazo, deduzidos do
valor dos investimentos, cujo resultado é semelhante.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{plot} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução do capital operacional líquido da empresa durante os períodos. Se for
\code{FALSE} ou \code{F}, o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador;
\item \strong{Índice} o capital operacional líquido dos periodos informados;
\item \strong{Análise Vertical} Análise Vertical das contas informadas no item 2.
Contas de resultado terão suas análises verticais em relação à receita total e
contas patrimoniais terão suas análises verticais em relação ao ativo total;
\item \strong{Análise Horizontal} Análise Horizontal das contas informadas no item 2.
}

Todos os itens da lista são bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de análise de dados.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{ind_capitalOperLiquido}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_capitalOperLiquido(
 indicador = "Liq Seca",
 periodo = 2021:2022,
 patLiq = c(150,200),
 investimentos = c(5,3),
 dividasCP = c(30, 40),
 dividasLP = c(60, 90),
 atvTotal = c(500,400),
 plot = TRUE,
 relatorio = TRUE
 )

}
