% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_liqCorrente.R
\name{ind_liqCorrente}
\alias{ind_liqCorrente}
\title{Índice de Liquidez Corrente}
\usage{
ind_liqCorrente(
  indicador = "Liq Corrente",
  periodo = 2019:2020,
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  atvTotal = c(900, 800),
  plot = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{cxEquiv}{Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)}

\item{estoque}{Vetor com os valores do estoque (Ativo Circulante)}

\item{ctaRecCP}{Vetor com os valores do contas a receber (Ativo Circulante)}

\item{outAtvCirc}{Vetor com os valores de outros ativos circulantes (Ativo Circulante)}

\item{fornec}{Vetor com os valores de fornecedores (Passivo Circulante)}

\item{dividasCP}{Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)}

\item{outPasCirc}{Vetor com os valores de outros passivos circulantes (Passivo Circulante)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}
}
\description{
Índice de Liquidez Corrente
}
\details{
Liquidez corrente é uma medida financeira que indica a capacidade de uma
empresa de cumprir suas obrigações de curto prazo. Calculada
dividindo-se o total de ativos circulantes pelo total de passivos
circulantes, a liquidez corrente é expressa em forma de números, sendo
valores acima de 1 indicativos de uma posição financeira saudável.
É representado pela fórmula:

\deqn{\frac{AC}{PC}}
em que: \strong{AC} é o Ativo Circulante e \strong{PC} é o Passivo Circulante

Para melhorar o processo de análise, os valores de Ativos Circulante e Passivo
Circulante foram desmembrados, respectivamente para
\eqn{AC = cxEquiv + estoque + ctaRecCP + outAtvCirc} e para
\eqn{PC = fornec + dividasCP + outPasCirc}. O item detalhes (details) apresenta
a descrição de cada conta.

Assim, tem-se que a liquidez corrente corresponde a:

A equação não contempla todas as contas do ativo circulante com exceção da
conta estoque.

A liquidez corrente pode ser útil para gestores financeiros, pois permite
avaliar a necessidade de recursos de curto prazo, como empréstimos bancários ou
financiamentos, para cobrir despesas operacionais
}
\examples{

suppressMessages(suppressWarnings(library(cntdd)))
suppressMessages(suppressWarnings(library(dplyr)))
suppressMessages(suppressWarnings(library(tidyr)))
suppressMessages(suppressWarnings(library(ggplot2)))

## Usando Vetores

ind_liqCorrente(
indicador  = "Liquidez Corrente",
periodo    = 2018:2020,
cxEquiv    = c(500,300,400),
estoque    = c(2000,3000,4000),
ctaRecCP   = c(2500, 5000, 2800),
outAtvCirc = c(20, 35, 80),
fornec     = c(1200, 1400, 1600),
dividasCP  = c(500, 200, 750),
outPasCirc = c(30, 20, 40),
atvTotal   = c(10000, 12000, 11000),
plot = T)

## Usando um data.frame
## Todos as variáveis devem ser do tipo numérico e os dados faltantes devem ser
## substituidos por zero.

dadosAlpha <- dt_contabil \%>\% filter(empresa == "alpha")

ind_liqCorrente(
indicador  = "Liquidez Corrente",
periodo    = dadosAlpha$ano,
cxEquiv    = dadosAlpha$cxEquiv,
estoque    = dadosAlpha$estoque,
ctaRecCP   = dadosAlpha$ctaRecCP,
outAtvCirc = dadosAlpha$outAtvCirc,
fornec     = dadosAlpha$fornec,
dividasCP  = dadosAlpha$dividasCP,
outPasCirc = dadosAlpha$outPasCirc,
atvTotal   = dadosAlpha$atvTotal,
plot = F)

}
