% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_liqGeral.R
\name{ind_liqGeral}
\alias{ind_liqGeral}
\title{Índice de Liquidez Geral}
\usage{
ind_liqGeral(
  indicador = "Liquidez Geral",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  atvRealizLP = c(2, 4),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  pasNaoCirc = c(6, 4),
  atvTotal = c(900, 800),
  relatorio = T,
  titulo = "Evolucao da Liquidez Geral",
  subtitulo = "",
  rodape = paste0("@", year(Sys.Date()), " contabiliDados"),
  corFundo = "orange",
  corLinhaTendencia = "orange",
  tamanhoValores = 6,
  tamanhoTempo = 10,
  tamanhoVariavel = 4,
  tamanhoTitulo = 14,
  tamanhoSubTitulo = 10,
  tamanhoRodape = 8,
  corRodape = "gray"
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{cxEquiv}{Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)}

\item{estoque}{Vetor com os valores do estoque (Ativo Circulante)}

\item{ctaRecCP}{Vetor com os valores do contas a receber (Ativo Circulante)}

\item{outAtvCirc}{Vetor com os valores de outros ativos circulantes (Ativo Circulante)}

\item{atvRealizLP}{Vetor com os valores do ativo realizável a longo prazo (Ativo Não Circulante)}

\item{fornec}{Vetor com os valores de fornecedores (Passivo Circulante)}

\item{dividasCP}{Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)}

\item{outPasCirc}{Vetor com os valores de outros passivos circulantes (Passivo Circulante)}

\item{pasNaoCirc}{Vetor com os valores do passivo não circulante}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{relatorio}{Se \code{TRUE}, Mostra relatório do indicador. Se \code{FALSE}, mostra
apenas o vetor com resultados do indicador (TRUE/FALSE)}

\item{titulo}{Título do gráfico}

\item{subtitulo}{Subtítulo do gráfico}

\item{rodape}{Rodapé do gráfico}

\item{corFundo}{Cor de fundo para os valores (Padrão: Laranja)}

\item{corLinhaTendencia}{Cor da linha de tendência entre os valores (Padrão: Laranja)}

\item{tamanhoValores}{Tamanho da fonte dos valores apresentados (Padrão: 6)}

\item{tamanhoTempo}{Tamanho da fonte dos rótulos relativo aos períodos de tempo (Padrão: 10)}

\item{tamanhoVariavel}{Tamanho da fonte do texto relativo à variável analisada (Padrão: 4)}

\item{tamanhoTitulo}{Tamanho da fonte do título (Padrão: 14)}

\item{tamanhoSubTitulo}{Tamanho da fonte do subtítulo (Padrão: 10)}

\item{tamanhoRodape}{Tamanho da fonte do rodapé (Padrão: 8)}

\item{corRodape}{Cor da fonte do rodapé (Padrão: Cinza)}
}
\description{
Essa função calcula o Índice de Liquidez Geral baseado em vetores relativos às
contas de ativo circulante / não circulante  e passivo circulante / não circulante.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{relatorio} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução do ativo cíclico da empresa durante os períodos. Se for \code{FALSE} ou \code{F},
o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador;
\item \strong{Índice} o índice de Liquidez Geral dos períodos informados;
\item \strong{Análise Vertical} Análise Vertical das contas informadas no item 2.
Contas de resultado terão suas análises verticais em relação à receita total e
contas patrimoniais terão suas análises verticais em relação ao ativo total;
\item \strong{Análise Horizontal} Análise Horizontal das contas informadas no item 2.
}

Todos os itens da lista são bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de análise de dados.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.com.br/}.
Ao acessar, fazer busca pelo nome da função \code{ind_liqGeral}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_liqGeral(
 indicador = "Liquidez Geral",
 periodo = 2021:2022,
 cxEquiv = c(8,10),
 estoque = c(150,200),
 ctaRecCP = c(400, 300),
 outAtvCirc = c(1, 3),
 atvRealizLP = c(2, 4),
 fornec = c(50, 20),
 dividasCP = c(30, 40),
 outPasCirc = c(10, 8),
 pasNaoCirc = c(6, 4),
 atvTotal = c(900,800),
 relatorio = TRUE,
 titulo = "Evolucao da Liquidez Geral",
 subtitulo = "",
 rodape = "",
 corFundo = "orange",
 corLinhaTendencia = "orange",
 tamanhoValores = 6,
 tamanhoTempo = 10,
 tamanhoVariavel = 4,
 tamanhoTitulo = 14,
 tamanhoSubTitulo = 10,
 tamanhoRodape = 8,
 corRodape = "gray"
 )

}
