% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_liqSeca.R
\name{ind_liqSeca}
\alias{ind_liqSeca}
\title{Índice de Liquidez Seca}
\usage{
ind_liqSeca(
  indicador = "Liq Seca",
  periodo = 2019:2020,
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  atvTotal = c(900, 800),
  plot = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{cxEquiv}{Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)}

\item{estoque}{Vetor com os valores do estoque (Ativo Circulante)}

\item{ctaRecCP}{Vetor com os valores do contas a receber (Ativo Circulante)}

\item{outAtvCirc}{Vetor com os valores de outros ativos circulantes (Ativo Circulante)}

\item{fornec}{Vetor com os valores de fornecedores (Passivo Circulante)}

\item{dividasCP}{Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)}

\item{outPasCirc}{Vetor com os valores de outros passivos circulantes (Passivo Circulante)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}
}
\description{
Índice de Liquidez Seca
}
\details{
O índice de Liquidez Seca revela a capacidade de pagamento de dívidas de curto
prazo, considerando que a empresa não consiga reverter o valor de seus estoques
em dinheiro. Relaciona, portanto, quanto uma empresa possui de ativo
(circulante), deduzido do valor de seus estoques, com o total de seu passivo
circulante. É representado pela fórmula:

\deqn{\frac{AC - Est}{PC}}
em que: \strong{AC} é o Ativo Circulante, \strong{Est} é o estoque e \strong{PC} é o Passivo Circulante

Para melhorar o processo de análise, os valores de Ativos Circulante e Passivo
Circulante foram desmembrados, respectivamente para
\eqn{AC = cxEquiv + estoque + ctaRecCP + outAtvCirc} e para
\eqn{PC = fornec + dividasCP + outPasCirc}. O item detalhes (details) apresenta
a descrição de cada conta.

Assim, tem-se que a liquidez seca corresponde a:

\deqn{\frac{(cxEquiv + ctaRecCP + outAtvCirc)}{(fornec + dividasCP + outPasCirc)}}

A equação não contempla todas as contas do ativo circulante com exceção da
conta estoque.

Indicamos essa análise quando a empresa apresentar dificuldade com o giro
de seus estoques. Em codições normais, não haveria motivos para a empresa não
conseguir rotacionar seus estoques. Quando há indicativos de que a rotação
dos estoques ficará prejudicada, esse indicador poderá representar melhor
a capacidade das empress em quitar suas dívidas de curto prazo.
}
\examples{

suppressMessages(suppressWarnings(library(cntdd)))
suppressMessages(suppressWarnings(library(dplyr)))
suppressMessages(suppressWarnings(library(tidyr)))
suppressMessages(suppressWarnings(library(ggplot2)))

## Usando Vetores

ind_liqSeca(
indicador  = "Liquidez Seca",
periodo    = 2018:2020,
cxEquiv    = c(500,300,400),
estoque    = c(2000,3000,4000),
ctaRecCP   = c(2500, 5000, 2800),
outAtvCirc = c(20, 35, 80),
fornec     = c(1200, 1400, 1600),
dividasCP  = c(500, 200, 750),
outPasCirc = c(30, 20, 40),
atvTotal   = c(10000, 12000, 11000),
plot = T)

## Usando um data.frame
## Todos as variáveis devem ser do tipo numérico e os dados faltantes devem ser
## substituidos por zero.

dadosAlpha <- dt_contabil \%>\% filter(empresa == "alpha")

ind_liqSeca(
indicador  = "Liquidez Seca",
periodo    = dadosAlpha$ano,
cxEquiv    = dadosAlpha$cxEquiv,
estoque    = dadosAlpha$estoque,
ctaRecCP   = dadosAlpha$ctaRecCP,
outAtvCirc = dadosAlpha$outAtvCirc,
fornec     = dadosAlpha$fornec,
dividasCP  = dadosAlpha$dividasCP,
outPasCirc = dadosAlpha$outPasCirc,
atvTotal   = dadosAlpha$atvTotal,
plot = F)

}
