% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_nopat.R
\name{ind_nopat}
\alias{ind_nopat}
\title{Lucro Operacional Liquido após os impostos (NOPAT)}
\usage{
ind_nopat(
  indicador = "nopat",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  ebit = c(150, 200),
  receitaLiquida = c(290, 230),
  ir = 0.34,
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numerico indicando o periodo da analise}

\item{ebit}{Vetor com os valores do Lucro antes dos Juros e Impostos (EBIT) da empresa}

\item{receitaLiquida}{Vetor com os valores da receita liquida da empresa}

\item{ir}{Vetor com o percentual do imposto de renda (padrao: 34\%)}

\item{plot}{Mostra grafico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatorio do indicador. Se \code{FALSE}, mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa funcao calcula o Lucro Operacional Liquido após os impostos (NOPAT -
Net operating profit after tax).
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Grafico} se o parametro \code{plot} for \code{TRUE} ou \code{T}, mostra um grafico com a
evolucao do NOPAT da empresa durante os periodos. Se for \code{FALSE} ou \code{F},
o grafico nao e apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
calculo do indicador;
\item \strong{Indice} o lucro operacional liquido dos periodos informados;
\item \strong{Analise Vertical} Analise Vertical das contas informadas no item 1. Contas de resultado
terao sua analise vertical em relacao a receita total e contas patrimoniais terao
sua analise vertical em relacao ao ativo total;
\item \strong{Analise Horizontal} Analise Horizontal das contas informadas no item 1.
}

Todos os itens da lista sao bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de analise de dados.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_roa}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
