% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_wacc.R
\name{ind_wacc}
\alias{ind_wacc}
\title{Custo Medio de Capital Ponderado (WACC ou CMPC)}
\usage{
ind_wacc(
  indicador = "WACC",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  empFinCP = c(200, 300),
  empFinLP = c(100, 200),
  patLiq = c(400, 600),
  despJuros = c(30, 40),
  receita = c(1000, 1200),
  atvTotal = c(4000, 5000),
  ir = 0.34,
  capm = T,
  ke = 0.15,
  beta = 1.3,
  rf = 0.07,
  rm = 0.13,
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numerico indicando o periodo da analise}

\item{empFinCP}{Vetor com os valores de emprestimos e financiamentos de curto prazo (Passivo Circulante)}

\item{empFinLP}{Vetor com os valores de emprestimos e financiamentos de longo prazo (Passivo Nao Circulante)}

\item{patLiq}{Vetor com os valores do Patrimônio Liquido}

\item{despJuros}{Vetor com os valores das despesas com juros (DRE/Notas Explicativas)}

\item{receita}{Vetor com os valores da receita liquida (Demonstracao de Resultado - DRE)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{ir}{Valor relativo a aliquota do Imposto de Renda}

\item{capm}{Se \code{TRUE}, calcula \code{ke} pela fórmula \code{rf + beta * (rm - rf)}. Se \code{FALSE}, calcula pelo parametro ke}

\item{ke}{Valor relativo ao custo de capital do sócio}

\item{beta}{Valor relativo ao beta (medida de volatilidade ou risco sistematico) de um titulo ou portifólio comparado ao mercado como um todo}

\item{rf}{Valor relativo a taxa de juros livre de risco ou risco baixo}

\item{rm}{Valor relativo a taxa de remuneracao paga pelo mercado}

\item{plot}{Se \code{TRUE}, Mostra grafico do indicador. Se \code{FALSE}, nao mostra o grafico (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatorio do indicador. Se \code{FALSE}, mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa funcao calcula o Custo Medio de Capital Ponderado baseado em vetores relativos ao
custo da divida, peso da divida e o beneficio fiscal do imposto de renda,
custo de capital do sócio, peso do capital do sócio.
}
\details{
Apresenta como resultado uma #' lista com 5 itens:
\enumerate{
\item \strong{Grafico} se o parametro \code{plot} for \code{TRUE} ou \code{T}, mostra um grafico com a
evolucao do Custo Medio de Capital Ponderado pelos valores investidos pelos
socios e banco. Se for \code{FALSE} ou \code{F}, o grafico nao e apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
calculo do indicador;
\item \strong{Indice} o indice de Custo Medio de Capital Ponderado dos periodos informados;
\item \strong{Analise Vertical} Analise Vertical das contas informadas no item 1. Conta de resultado
terao sua analise vertical em relacao a receita total e contas patrimoniais terao
sua analise vertical em relacao ao ativo total;
\item \strong{Analise Horizontal} Analise Horizontal das contas informadas no item 1.
}

Todos os itens da lista sao bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de analise de dados.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_roa}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
