% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meses.R
\name{meses}
\alias{meses}
\title{Meses do ano}
\usage{
meses()
}
\description{
Esta função gera um data frame com os meses do anos em diversos formatos
e informações adicionais como bimestre, trimestre, quadrimestre e semestre
}
