% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_correl.R
\name{stat_correl}
\alias{stat_correl}
\title{Correlação de Pearson e Spearman}
\usage{
stat_correl(
  dt = cntdd::dt_contabil[4:8],
  pearsonLower = T,
  ommitLower = F,
  ommitUpper = F,
  digits = 2
)
}
\arguments{
\item{dt}{Um data.frame}

\item{pearsonLower}{Se \code{TRUE} ou \code{T}, a correlação de Pearson é disposta na diagonal inferior da Matriz e a correlação de Spearman na diagonal superior. Se \code{FALSE} OU \code{F}, a disposição das correlações é inversa. (Padrão: \code{T})}

\item{ommitLower}{Se \code{TRUE} ou \code{T}, a diagonal INFERIOR é omitida. Se \code{FALSE} OU \code{F}, é mostrada. (Padrão: \code{F})}

\item{ommitUpper}{Se \code{TRUE} ou \code{T}, a diagonal SUPERIOR é omitida. Se \code{FALSE} OU \code{F}, é mostrada. (Padrão: \code{F})}

\item{digits}{Número de dígitos da correlação (Padrão: \code{2})}
}
\description{
Gera matriz de correlação de Pearson e/ou Spearman com suas respectivas
significâncias estatísticas
}
\examples{

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
`cntdd` do Blog do Projeto contabiliDados: \href{http://contabilidados.com.br}{(Acesse Aqui)}. Ao acessar, fazer busca
pelo nome da função `ind_liqSeca`

Contatos pelo email do Projeto contabiliDados:
Email: \email{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \href{https://www.instagram.com/contabilidados/}{@contabilidados}

}
