% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_meanTest.R
\name{stat_meanTest}
\alias{stat_meanTest}
\title{Teste de Medias ou Medianas}
\usage{
stat_meanTest(value_Y, name_Y = "Y", value_X, name_X = "X", pvalor = 0.05)
}
\arguments{
\item{value_Y}{Um vetor numerico representando o primeiro grupo}

\item{name_Y}{Um nome para o vetor numerico que representa o primeiro grupo}

\item{value_X}{Um vetor numerico representando o segundo grupo}

\item{name_X}{Um nome para o vetor numerico que representa o segundo grupo}

\item{pvalor}{O valor cretico a ser considerado nos testes. O padrao e 5\%}
}
\value{
Um data.frame com a analise sobre o teste de media dos dois grupos de valores.
}
\description{
Esta funcao realiza o teste de diferenca de medias ou mediana de dois vetores numericos
e apresenta os resultados em uma tabela.
}
\details{
A escolha entre o teste de media ou media e realizada de forma automatica, de acordo com a distribuicao
das variaveis numericas informadas.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
