% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testaIndependencia.R
\name{testaIndependencia}
\alias{testaIndependencia}
\title{Teste de Independência}
\usage{
testaIndependencia(bd, correcao = T, ...)
}
\arguments{
\item{bd}{Um data.frame com as variaveis categóricas}

\item{correcao}{Se \code{TRUE}, fará a correção dos erros}

\item{...}{a lista de variaveis categóricas de interesse}
}
\description{
Teste de Independência
}
\details{
Essa função faz o teste de independência entre variáveis categóricas e retorna
uma lista com a análise correspondente.
}
\examples{

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
`cntdd` do Blog do Projeto contabiliDados: \href{http://contabilidados.com.br}{(Acesse Aqui)}. Ao acessar, fazer busca
pelo nome da função `ind_liqSeca`

Contatos pelo email do Projeto contabiliDados:
Email: \email{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \href{https://www.instagram.com/contabilidados/}{@contabilidados}

}
