% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_ajustaContaInflacao.R
\name{utl_ajustaContaInflacao}
\alias{utl_ajustaContaInflacao}
\title{Ajuste de Valores pela Inflação}
\usage{
utl_ajustaContaInflacao(
  periodo = 2019:2023,
  vrConta = c(100, 110, 120, 130, 140),
  indiceInflacao = c(4.31, 4.52, 10.06, 5.79, 4.62),
  periodoBase = 2023,
  apenasVetor = T
)
}
\arguments{
\item{periodo}{Vetor contendo o período dos dados}

\item{vrConta}{Vetor contendo o valor da conta que pretende atualizar pela inflação}

\item{indiceInflacao}{Valor percentual do indice de inflacao do período.
Se o valor for 2,5\%, por exemplo, informar 2.5}

\item{periodoBase}{Período específico para o qual deseja atualizar os valores}

\item{apenasVetor}{Se \code{TRUE} ou \code{T} (padrão), o resultado é um vetor com os valores
atualizados. Se \code{FALSE} ou \code{F}, o resultado é um data.frame com todos os valores
envolvidos no cálculo do ajuste à inflação}
}
\description{
Faz o ajuste do valor de uma conta pelo valor do índice de inflação informado.
}
\details{
Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{utl_ajustaContaInflacao}.

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

utl_ajustaContaInflacao(
periodo = 2019:2023,
vrConta = c(100, 110, 120, 130, 140),
indiceInflacao = c(4.31, 4.52, 10.06, 5.79, 4.62),
periodoBase = 2023, apenasVetor = TRUE)

}
