% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_baixaBCB.R
\name{utl_baixaBCB}
\alias{utl_baixaBCB}
\title{Baixa dados Banco Central}
\usage{
utl_baixaBCB(codSerie = 433, inicio = "01/01/2003", fim = "31/12/2017")
}
\arguments{
\item{codSerie}{Valor do código da série no Banco Central}

\item{inicio}{String contendo a data de início da série no formato dd/mm/aaaa}

\item{fim}{String contendo a data final da série no formato dd/mm/aaaa}
}
\description{
Baixa dados da API do Banco Central
}
\details{
Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{utl_baixaBCB}.

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

utl_baixaBCB(codSerie = 433, inicio = "01/01/2003", fim = "31/12/2017")

}
