% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_consultarCNPJ.R
\name{utl_consultarCNPJ}
\alias{utl_consultarCNPJ}
\title{Consulta de informações de CNPJ via API "minhareceita.org}
\usage{
utl_consultarCNPJ(cnpj = "24529265000140")
}
\arguments{
\item{cnpj}{Um vetor de CNPJs como caracteres.}
}
\value{
Um DataFrame com informações detalhadas das empresas consultadas.
}
\description{
Esta função consulta informações detalhadas de empresas a partir de uma lista de CNPJs,
utilizando a API "minhareceita.org".
}
\details{
A função retorna um DataFrame contendo informações como UF, CEP, razão social,
entre outros. Inclui subfunções para obter dados gerais, CNAEs  secundários e
dados societários.
}
\examples{
library(cntdd)

utl_consultarCNPJ <- function(cnpj = "24529265000140")

}
\seealso{
\code{\link[=jsonlite]{jsonlite()}}, \code{\link[=dplyr]{dplyr()}}

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{utl_consultarCNPJ}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
