% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_createGroup.R
\name{utl_createGroup}
\alias{utl_createGroup}
\title{Cria grupos}
\usage{
utl_createGroup(bd, col_Value, col_Grp, n_grp = 2)
}
\arguments{
\item{bd}{Um data.frame}

\item{col_Value}{String informando o nome da coluna referente à variável númerica sobre a qual se pretende criar os grupos}

\item{col_Grp}{String com o nome da nova coluna a ser criada com os grupos}

\item{n_grp}{Número de grupos, se pretende criar grupos com o mesmo número de observações ou um vetor contendo os percentis para cada grupo. Padrão igual 2.}

\item{lbl_grp}{String com o prefixo posto no início do nome do de cada grupo. Padrão igual "grp"}
}
\description{
Cria grupos baseados nos percentise de uma variável númerica
}
\examples{

## Grupo 1 com os primeiros 20\% das observações de menor valor, Grupo 2 com as observações de
## 20\% a 80\% e Grupo 3 com os 20\% das observações de maior valor
utl_createGroup(mtcars, "mpg", "grp.mpg", c(0, .2, .8, 1))

## Grupo 1 com metade das observações e Grupo 2 com a metade restante
utl_createGroup(mtcars, "mpg", "grp.mpg", 2)


}
