% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_qdeNAcols.R
\name{utl_qdeNAcols}
\alias{utl_qdeNAcols}
\title{Contador de valores nulos em colunas}
\usage{
utl_qdeNAcols(dt, ...)
}
\arguments{
\item{dt}{Um data.frame}

\item{...}{Um vetor que irá receber os nomes das colunas que terão os valores nulos contabilizados}
}
\description{
Faz a contagem de valores nulos/inexistentes em cada coluna.
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
