% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robError.R
\name{utl_robErro}
\alias{utl_robErro}
\title{Erros robustos}
\usage{
utl_robErro(reg, tipo = "HC1", HAC = F)
}
\arguments{
\item{reg}{Um objeto do tipo "lm"}

\item{tipo}{String informando qual a correção da heterocedasticidade (HC1, HC2...)}

\item{HAC}{Se FALSE(default), os erros serao robustos a heterocedasticidade. Se TRUE, serao robustos a heterocedasticidade e a autocorrelacao.}
}
\description{
Estimacao consistente com a heterocedasticidade e/ou autocorrelacao da matriz
de covariancia das estimativas de coeficientes em modelos de regressao.
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\seealso{
\code{\link[=cut]{cut()}}, \code{\link[=quantile]{quantile()}}, \code{\link[=vcovHAC]{vcovHAC()}}, \code{\link[=vcovHC]{vcovHC()}}
}
