% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_t_starSig.R
\name{utl_tStarSig}
\alias{utl_tStarSig}
\title{Formata significancia}
\usage{
utl_tStarSig(
  sig,
  t = NULL,
  nDigits = 2,
  sigValue = c(0.01, 0.05, 0.1),
  sigSimbolo = c("***", "**", "*")
)
}
\arguments{
\item{sig}{o P-Value}

\item{t}{a Estatística t}

\item{nDigits}{Numero de Digitos}

\item{sigValue}{Valores da significancia (Padrao: 1\%, 5\% e 10\%)}

\item{sigSimbolo}{Simbolos usados para indicar as significancias (Padrao: 1\%, 5\% e 10\%)}
}
\description{
Essa funcao formata a estatistica t e o nivel de significancia. Seu resultado
pode ser uma Estatistica t com o numero de casas decimais informados com o
respectivo p-valor identificado por asteriscos.
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
