% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtContabil.R
\docType{data}
\name{dt_contabil}
\alias{dt_contabil}
\title{Dados Contábeis}
\format{
Data.frame com 40 linhas e 35 colunas.
}
\usage{
dt_contabil
}
\description{
Dados contábeis de empresas fictícias relacionados ao Balanco Patrimonial (BP),
Demonstração do Resultado do Exercício (DRE) e Demonstração dos Fluxos de
Caixa (DFC).
}
\details{
Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.com.br/}.
Ao acessar, fazer busca pelo nome da funcao \code{dt_contabil}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
