% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtContabil.R
\docType{data}
\name{dt_contabil}
\alias{dt_contabil}
\title{Dados Contábeis}
\format{
An object of class \code{data.frame} with 40 rows and 35 columns.
}
\usage{
dt_contabil
}
\description{
Dados contabeis de empresas fictícias relacionados ao Balanço Patrimonial (BP),
Demonstração do Resultado do Exercício (DRE) e Demonstração dos Fluxos de
Caixa (DFC).
}
\details{
Data.frame com 40 linhas e 35 colunas.
}
\keyword{datasets}
