% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtContabil.R
\docType{data}
\name{dt_contabil}
\alias{dt_contabil}
\title{Dados Contabeis}
\format{
Data.frame com 40 linhas e 35 colunas.
}
\usage{
dt_contabil
}
\description{
Dados contabeis de empresas ficticias relacionados ao Balanco Patrimonial (BP),
Demonstracao do Resultado do Exercicio (DRE) e Demonstracao dos Fluxos de
Caixa (DFC).
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{dt_contabil}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
