% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmeses.R
\docType{data}
\name{dt_meses}
\alias{dt_meses}
\title{Meses do ano}
\format{
Data.frame com 12 linhas e 10 colunas.
}
\usage{
dt_meses
}
\description{
Variacao de apresentacao dos dados de meses (textual ou numerico). Tambem sao
incorporadas informacoes sobre bimestre, trimestre, quadrimestre e semestre
relativos a cada mes do ano.
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{dt_meses}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
