% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmeses.R
\docType{data}
\name{dt_meses}
\alias{dt_meses}
\title{Meses do ano}
\format{
Data.frame com 12 linhas e 10 colunas.
}
\usage{
dt_meses
}
\description{
Variação de apresentação dos dados de meses (textual ou numérico). Também são
incorporadas informações sobre bimestre, trimestre, quadrimestre e semestre
relativos a cada mês do ano.
}
\details{
Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub}.
Ao acessar, fazer busca pelo nome da função \code{dt_meses}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
