% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtUFregiao.R
\docType{data}
\name{dt_ufRegiao}
\alias{dt_ufRegiao}
\title{Dados das Unidades Federativas}
\format{
An object of class \code{data.frame} with 27 rows and 4 columns.
}
\usage{
dt_ufRegiao
}
\description{
Dados dos estados brasileiros com suas respectivas siglas de Unidade
Federativa (UF), capital e região. Útil para associação com outras bases.
}
\details{
Data.frame com 27 observações (linhas) e 4 variáveis (colunas).

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub}.
Ao acessar, fazer busca pelo nome da função \code{dt_ufRegiao}.

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\keyword{datasets}
