% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_analiseCiclo.R
\name{ind_analiseCiclo}
\alias{ind_analiseCiclo}
\title{Análise Dupont (ROE desmembrado)}
\usage{
ind_analiseCiclo(
  intervalo = 2009:2011,
  patrimonioLiquido = c(200, 150, 175),
  ativoTotal = c(300, 500, 600),
  receita = c(500, 450, 400),
  lucroLiquido = c(8, 10, 14),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{intervalo}{Vetor numérico indicando o período da análise}

\item{patrimonioLiquido}{Vetor com os valores do patrimônio líquido da empresa}

\item{ativoTotal}{Vetor com os valores do ativo total da empresa}

\item{receita}{Vetor com os valores da receita líquida da empresa}

\item{lucroLiquido}{Vetor com os valores do lucro líquido da empresa}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatório do indicador. Se \code{FALSE}, mostra
apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa função calcula o Índice de Retorno sobre o Patrimônio Líquido, baseado
em vetores relativos as contas de Lucro Liquido, receita, Ativo Total e
endividamento.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{plot} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução do ROE da empresa durante os períodos. Se for \code{FALSE} ou \code{F},
o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador;
\item \strong{Índice} o índice de Retorno sobre o patrimônio líquido dos períodos informados;
\item \strong{Análise Vertical} Análise Vertical das contas informadas no item 2.
Contas de resultado terão suas análises verticais em relação à receita total e
contas patrimoniais terão suas análises verticais em relação ao ativo total;
\item \strong{Análise Horizontal} Análise Horizontal das contas informadas no item 2.
}

Todos os itens da lista são bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de análise de dados.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{ind_dupont}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_analiseDupont(
 intervalo = 2009:2011,
 patrimonioLiquido = c(200, 150, 175),
 ativoTotal = c(300, 500, 600),
 receita = c(500, 450, 400),
 lucroLiquido = c(8, 10, 14),
 plot = TRUE,
 relatorio = TRUE
 )

}
\seealso{
\code{\link[=ind_roe]{ind_roe()}}, \code{\link[=ind_roa]{ind_roa()}}, \code{\link[=ind_nivelEndividamento]{ind_nivelEndividamento()}}, \code{\link[=ind_giro]{ind_giro()}}, \code{\link[=ind_ros]{ind_ros()}}, \code{\link[=ind_pmre]{ind_pmre()}}
}
