% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_capitalOperLiquido.R
\name{ind_capitalOperLiquido}
\alias{ind_capitalOperLiquido}
\title{Capital Operacional Liquido (COL)}
\usage{
ind_capitalOperLiquido(
  indicador = "Liq Seca",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  patLiq = c(150, 200),
  investimentos = c(5, 3),
  dividasCP = c(30, 40),
  dividasLP = c(60, 90),
  atvTotal = c(500, 400),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numerico indicando o periodo da analise}

\item{patLiq}{Vetor com os valores do patrimonio liquido da empresa}

\item{investimentos}{Vetor com os valores dos investimentos da empresa}

\item{dividasCP}{Vetor com os valores de emprestimos e financiamentos de curto prazo (Passivo Circulante)}

\item{dividasLP}{Vetor com os valores de emprestimos e financiamentos de longo prazo (Passivo Nao Circulante)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra grafico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatorio do indicador. Se \code{FALSE}, mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa funcao calcula o Capital Operacional Liquido de forma simplificada, pois
nao calcula o ativo e passivo operacional, mas faz uma simplificacao pela
soma do patrimonio liquido com a divida de curto e longo prazo, deduzidos do
valor dos investimentos. Em linhas gerais, o resultado e semelhante.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Grafico} se o parametro \code{plot} for \code{TRUE} ou \code{T}, mostra um grafico com a
evolucao capital operacional líquido da empresa durante os periodos. Se for \code{FALSE} ou \code{F},
o grafico nao e apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
calculo do indicador;
\item \strong{Indice} o capital operacional liquido dos periodos informados;
\item \strong{Analise Vertical} Analise Vertical das contas informadas no item 1. Contas de resultado
terao sua analise vertical em relacao a receita total e contas patrimoniais terao
sua analise vertical em relacao ao ativo total;
\item \strong{Analise Horizontal} Analise Horizontal das contas informadas no item 1.
}

Todos os itens da lista sao bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de analise de dados.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_roa}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
