% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dupont.R
\name{ind_dupont}
\alias{ind_dupont}
\title{Analise Dupont (ROE desmembrado)}
\usage{
ind_dupont(
  indicador = "ROE_Dupont",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  lucLiq = c(80, 100),
  patLiq = c(180, 190),
  ativoTotal = c(200, 300),
  receitaTotal = c(300, 500),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numerico indicando o periodo da analise}

\item{lucLiq}{Vetor com os valores do lucro liquido da organizacao}

\item{patLiq}{vetor com os valores do patrimonio liquido da organizacao}

\item{ativoTotal}{Vetor com os valores do ativo total da empresa}

\item{receitaTotal}{Vetor com os valores da receita total da empresa}

\item{plot}{Mostra grafico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatorio do indicador. Se \code{FALSE}, mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa funcao calcula o Indice de Retorno sobre o Ativo baseado em vetores relativos as
contas de Lucro Liquido e Ativo Total.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Grafico} se o parametro \code{plot} for \code{TRUE} ou \code{T}, mostra um grafico com a
evolucao do ROE da empresa durante os periodos. Se for \code{FALSE} ou \code{F},
o grafico nao e apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
calculo do indicador;
\item \strong{Indice} o indice de Retorno sobre o patrimonio liquido dos periodos informados;
\item \strong{Analise Vertical} Analise Vertical das contas informadas no item 1. Conta de resultado
terao sua analise vertical em relacao a receita total e contas patrimoniais terao
sua analise vertical em relacao ao ativo total;
\item \strong{Analise Horizontal} Analise Horizontal das contas informadas no item 1.
}

Todos os itens da lista sao bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de analise de dados.

Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{ind_roa}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
