% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_liqCorrente.R
\name{ind_liqCorrente}
\alias{ind_liqCorrente}
\title{Índice de Liquidez Corrente}
\usage{
ind_liqCorrente(
  indicador = "Liq Corrente",
  periodo = 2019:2020,
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  atvTotal = c(900, 800),
  plot = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{cxEquiv}{Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)}

\item{estoque}{Vetor com os valores do estoque (Ativo Circulante)}

\item{ctaRecCP}{Vetor com os valores do contas a receber (Ativo Circulante)}

\item{outAtvCirc}{Vetor com os valores de outros ativos circulantes (Ativo Circulante)}

\item{fornec}{Vetor com os valores de fornecedores (Passivo Circulante)}

\item{dividasCP}{Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)}

\item{outPasCirc}{Vetor com os valores de outros passivos circulantes (Passivo Circulante)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}
}
\description{
Liquidez corrente é uma medida financeira que indica a capacidade de uma
empresa de cumprir suas obrigações de curto prazo. Calculada
dividindo-se o total de ativos circulantes pelo total de passivos
circulantes, a liquidez corrente é expressa em forma de números, sendo
valores acima de 1 indicativos de uma posição financeira saudável.

A liquidez corrente pode ser útil para gestores financeiros, pois permite
avaliar a necessidade de recursos de curto prazo, como empréstimos bancários ou
financiamentos, para cobrir despesas operacionais.
}
\examples{

Informações adicionais sobre como usar o pacote no menu cntdd
 do link: \href{http://contabilidados.com.br}{contabiliDados}.

Contatos pelo email do Projeto contabiliDados:
\email{contabilidados@ufersa.edu.br}


}
