% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_liqSeca.R
\name{ind_liqSeca}
\alias{ind_liqSeca}
\title{Índice de Liquidez Seca}
\usage{
ind_liqSeca(indicador, periodo, estoque, atvCirc, pasCirc, atvTotal)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{estoque}{Vetor com os valores do estoque}

\item{atvCirc}{Vetor com os valores do Ativo Circulante}

\item{pasCirc}{Vetor com os valores do Passivo Circulante}
}
\description{
Esse índice corresponde à capacidade de um vetor numérico
}
\examples{

## Usando Vetores
ind_liqSeca(indicador = "Liquidez Seca",
periodo = 2018:2020,
estoque = c(2000,3000,4000),
atvCirc = c(5000,6000,7000),
pasCirc = c(1000,2000,3000))

## Usando como relatório HTML
analise <-
ind_liqSeca(indicador = "Liquidez Seca",
periodo = 2018:2020,
estoque = c(2000,3000,4000),
atvCirc = c(5000,6000,7000),
pasCirc = c(1000,2000,3000))

analise$`Como Analisar` \%>\% sjPlot::tab_df(encoding = "Latim-1")

## Usando um data.frame

}
