% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_liqSeca.R
\name{ind_liqSeca}
\alias{ind_liqSeca}
\title{Índice de Liquidez Seca}
\usage{
ind_liqSeca(
  indicador = "Liq Seca",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  atvTotal = c(900, 800),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{cxEquiv}{Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)}

\item{estoque}{Vetor com os valores do estoque (Ativo Circulante)}

\item{ctaRecCP}{Vetor com os valores do contas a receber (Ativo Circulante)}

\item{outAtvCirc}{Vetor com os valores de outros ativos circulantes (Ativo Circulante)}

\item{fornec}{Vetor com os valores de fornecedores (Passivo Circulante)}

\item{dividasCP}{Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)}

\item{outPasCirc}{Vetor com os valores de outros passivos circulantes (Passivo Circulante)}

\item{atvTotal}{Vetor com os valores do Ativo Total}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatório do indicador. Se \code{FALSE},
mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa função calcula o Índice de Liquidez Seca baseado em vetores relativos as
contas de ativo circulante, estoques e passivo circulante.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{plot} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução da Liquidez Seca da empresa durante os períodos. Se for \code{FALSE} ou \code{F},
o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador;
\item \strong{Índice} o índice de Liquidez Seca dos períodos informados;
\item \strong{Análise Vertical} Análise Vertical das contas informadas no item 2.
Contas de resultado terão suas análises verticais em relação à receita total e
contas patrimoniais terão suas análises verticais em relação ao ativo total;
\item \strong{Análise Horizontal} Análise Horizontal das contas informadas no item 2.
}

Todos os itens da lista são bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de análise de dados.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{ind_liqSeca}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_liqSeca(
    indicador = "Liq Seca",
    periodo = 2021:2022,
    cxEquiv = c(8,10),
    estoque = c(150,200),
    ctaRecCP = c(400, 300),
    outAtvCirc = c(1, 3),
    fornec = c(50, 20),
    dividasCP = c(30, 40),
    outPasCirc = c(10, 8),
    atvTotal = c(900,800),
    plot = TRUE,
    relatorio = TRUE
)

}
