% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_roe.R
\name{ind_roe}
\alias{ind_roe}
\title{Índice de Retorno sobre o Patrimônio Líquido (ROE)}
\usage{
ind_roe(
  indicador = "Roe",
  periodo = 2019:2020,
  lucLiq = c(8, 10),
  patLiq = c(150, 200),
  receitaTotal = c(300, 500),
  passivoTotal = c(400, 500),
  plot = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{lucLiq}{Vetor com os valores do lucro líquido da empresa}

\item{patLiq}{Vetor com os valores do patrimônio líquido}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}
}
\description{
Essa função calcula o Índice de Retorno sobre o Patrimônio Líquido baseado em vetores relativos às
contas de lucro líquido e patrimônio líquido. Apresenta como resultado uma
lista com 5 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{plot} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução do ROE da empresa durante os períodos. Se for \code{FALSE} ou \code{F},
o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador;
\item \strong{Índice} o índice de Retorno sobre o Patrimônio Líquido dos períodos informados;
\item \strong{Análise Vertical} Análise Vertical das contas informadas no item 1;
\item \strong{Análise Horizontal} Análise Horizontal das contas informadas no item 1.
}

Todos os itens da lista são bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de análise de dados.
}
\examples{

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
`cntdd` do Blog do Projeto contabiliDados: \href{http://contabilidados.com.br}{(Acesse Aqui)}. Ao acessar, fazer busca
pelo nome da função `ind_roe`


Contatos pelo email do Projeto contabiliDados:
Email: \email{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \href{https://www.instagram.com/contabilidados/}{@contabilidados}


}
