% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ros.R
\name{ind_ros}
\alias{ind_ros}
\title{Índice de Retorno sobre as Vendas (ROS)}
\usage{
ind_ros(
  indicador = "ROS",
  periodo = (year(Sys.Date()) - 2):(year(Sys.Date()) - 1),
  lucro = c(8, 10),
  receitaTotal = c(300, 500),
  plot = T,
  relatorio = T
)
}
\arguments{
\item{indicador}{Um vetor tipo character com o nome do indicador}

\item{periodo}{Vetor numérico indicando o período da análise}

\item{lucro}{Vetor com os valores do lucro da empresa}

\item{receitaTotal}{Vetor com os valores da receita líquida da empresa}

\item{plot}{Mostra gráfico? (TRUE/FALSE)}

\item{relatorio}{Se \code{TRUE}, Mostra relatório do indicador. Se \code{FALSE},
mostra apenas o vetor com resultados do indicador (TRUE/FALSE)}
}
\description{
Essa função calcula o Índice de Retorno sobre as Vendas baseado em vetores relativos às
contas de Lucro e Receitas. O usuário definirá se usará lucro líquido ou
outra medida de lucro, conforme literatura de interesse. Da mesma forma,
escolherá entre receita líquida ou bruta. A literatura, especialmente sobre
análise Dupont, trata essa medida como \strong{margem}.
}
\details{
Apresenta como resultado uma lista com 5 itens:
\enumerate{
\item \strong{Gráfico} se o parâmetro \code{plot} for \code{TRUE} ou \code{T}, mostra um gráfico com a
evolução do ROS da empresa durante os períodos. Se for \code{FALSE} ou \code{F},
o gráfico não é apresentado;
\item \strong{Contas} que corresponde ao banco de dados com as contas informadas para
cálculo do indicador;
\item \strong{Índice} o ROS dos períodos informados;
\item \strong{Análise Vertical} Análise Vertical das contas informadas no item 2.
Contas de resultado terão suas analises verticais em relação à receita total e
contas patrimoniais terão suas analises verticais em relação ao ativo total;
\item \strong{Análise Horizontal} Análise Horizontal das contas informadas no item 2.
}

Todos os itens da lista são bancos de dados no formato tibble que podem ser
usados individualmente durante o processo de análise de dados.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{ind_ros}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

ind_ros(
 indicador = "ROS",
 periodo = 2021:2022,
 lucro = c(8,10),
 receitaTotal = c(300,500),
 plot = TRUE,
 relatorio = TRUE
 )

}
