% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_correl.R
\name{stat_correl}
\alias{stat_correl}
\title{Correlacao de Pearson e Spearman}
\usage{
stat_correl(
  dt = cntdd::dt_contabil[4:8],
  pearsonLower = T,
  ommitLower = F,
  ommitUpper = F,
  digits = 2,
  decimal = ","
)
}
\arguments{
\item{dt}{Um data.frame}

\item{pearsonLower}{Se \code{TRUE} ou \code{T}, a correlacao de Pearson e disposta na diagonal inferior da Matriz e a correlação de Spearman na diagonal superior. Se \code{FALSE} OU \code{F}, a disposicao das correlacoes e inversa. (Padrao: \code{T})}

\item{ommitLower}{Se \code{TRUE} ou \code{T}, a diagonal INFERIOR e omitida. Se \code{FALSE} OU \code{F}, e mostrada. (Padrao: \code{F})}

\item{ommitUpper}{Se \code{TRUE} ou \code{T}, a diagonal SUPERIOR e omitida. Se \code{FALSE} OU \code{F}, e mostrada. (Padrao: \code{F})}

\item{digits}{Numero de digitos da correlacao (Padrao: \code{2})}

\item{decimal}{Caracter para casa decimal (Padrao: \verb{,})}
}
\description{
Gera matriz de correlacao de Pearson e/ou Spearman com suas respectivas
significancias estatisticas
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
