% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_geomMean.R
\name{stat_geomMean}
\alias{stat_geomMean}
\title{Calcular a média geométrica}
\usage{
stat_geomMean(x, dec = 3)
}
\arguments{
\item{x}{Um vetor numérico}

\item{dec}{O número de casas decimais (padrão: 3)}
}
\description{
Esta função calculará a média geométrica de um vetor numérico
}
\examples{
library(cntdd)
stat_geomMean(1:6, 2) # Vetor

stat_geomMean(dt_contabil$atvTotal) # DataFrame

suppressMessages(suppressWarnings(library(dplyr)))
dt_contabil \%>\% summarise(estoqueMedio = stat_geomMean(estoque)) # Dplyr

}
