% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_geomMean.R
\name{stat_geomMean}
\alias{stat_geomMean}
\title{Media geometrica}
\usage{
stat_geomMean(x, dec = 3)
}
\arguments{
\item{x}{Um vetor numerico}

\item{dec}{O numero de casas decimais (padrao: 3)}
}
\description{
Esta funcao calculara a media geometrica de um vetor numerico.
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
