% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_meanTest.R
\name{stat_meanTest}
\alias{stat_meanTest}
\title{Teste de Médias ou Medianas}
\usage{
stat_meanTest(value_Y, name_Y = "Y", value_X, name_X = "X", pvalor = 0.05)
}
\arguments{
\item{value_Y}{Um vetor numérico representando o primeiro grupo}

\item{name_Y}{Um nome para o vetor numérico que representa o primeiro grupo}

\item{value_X}{Um vetor numérico representando o segundo grupo}

\item{name_X}{Um nome para o vetor numérico que representa o segundo grupo}

\item{pvalor}{O valor crítico a ser considerado nos testes. O padrão é 5\%}
}
\value{
Um data.frame com a análise sobre o teste de média/mediana dos dois grupos de valores.
}
\description{
Esta função realiza o teste de diferença de médias ou mediana de dois vetores numéricos
e apresenta os resultados em uma tabela.
}
\details{
A escolha entre o teste de média ou mediana é realizada de forma automática,
de acordo com a distribuição das variáveis numéricas informadas.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{stat_meanTest}.

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)
library(dplyr)
library(nortest)

dt <- na.omit(cntdd::dt_contabil)

dt <- cntdd::utl_createGroup(
 bd = dt, col_Value = "receita", col_Grp = "grp.Receita",
 n_grp = 2) \%>\%
 select(grp.Receita, caixaEquiv)

 grupo1 <- dt \%>\% filter(grp.Receita == "receita_1") \%>\% select(caixaEquiv) \%>\% pull()
 grupo2 <- dt \%>\% filter(grp.Receita == "receita_2") \%>\% select(caixaEquiv) \%>\% pull()

 stat_meanTest(
   value_Y = grupo1, name_Y = "Menor Receita",
   value_X = grupo2, name_X = "Maior Receita"
   )

}
