% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_testaIndependencia.R
\name{stat_testaIndependencia}
\alias{stat_testaIndependencia}
\title{Teste de Independencia}
\usage{
stat_testaIndependencia(bd, correcao = T, ...)
}
\arguments{
\item{bd}{Um data.frame com as variaveis categoricas}

\item{correcao}{Se \code{TRUE}, fara a correcao dos erros}

\item{...}{a lista de variaveis categoricas de interesse}
}
\description{
Essa funcao faz o teste de independencia entre variaveis categoricas e retorna
uma lista com a analise correspondente.
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
