% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_consultarCNPJ.R
\name{utl_consultarCNPJ}
\alias{utl_consultarCNPJ}
\title{Consulta de informações de CNPJ via API "minhareceita.org}
\usage{
utl_consultarCNPJ(cnpj = "24529265000140")
}
\arguments{
\item{cnpj}{Um vetor de CNPJs como caracteres.}
}
\value{
Um DataFrame com informações detalhadas das empresas consultadas.
}
\description{
Esta função consulta informações detalhadas de empresas a partir de uma lista de CNPJs,
utilizando a API "minhareceita.org". A função retorna um DataFrame contendo informações
como UF, CEP, razão social, entre outros. Inclui subfunções para obter dados gerais, CNAEs
secundários e dados societários.
}
\details{
A função depende das bibliotecas \code{jsonlite} e \code{tidyverse}. Certifique-se de que estão instaladas
e carregadas.
}
\examples{
cnpj <- c("08350241000172", "24529265000140")
df_resultado <- utl_consultarCNPJ(cnpj)

Função desenvolvida por Alexsandro Prado (alexsandro.prado@ufersa.edu.br) em colaboração com Kleber Formiga (mirandakf@ufersa.edu.br)
Informações adicionais sobre como usar o pacote, orientamos acessar o menu
`cntdd` do Blog do Projeto contabiliDados: \href{http://contabilidados.com.br}{(Acesse Aqui)}. Ao acessar, fazer busca
pelo nome da função `ind_liqCorrente`

Contatos pelo email do Projeto contabiliDados:
Email: \email{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \href{https://www.instagram.com/contabilidados/}{@contabilidados}


}
