% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_createGroup.R
\name{utl_createGroup}
\alias{utl_createGroup}
\title{Cria grupos}
\usage{
utl_createGroup(bd, col_Value, col_Grp, n_grp = 2)
}
\arguments{
\item{bd}{Um data.frame}

\item{col_Value}{String informando o nome da coluna referente a variavel numerica sobre a qual se pretende criar os grupos}

\item{col_Grp}{String com o nome da nova coluna a ser criada com os grupos}

\item{n_grp}{Número de grupos, se pretende criar grupos com o mesmo numero de observacoes ou um vetor contendo os percentis para cada grupo. Padrao igual 2.}
}
\description{
Cria grupos baseados nos percentis de uma variavel numerica
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\seealso{
\code{\link[=cut]{cut()}}, \code{\link[=quantile]{quantile()}}
}
