% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_limpaCNPJ.R
\name{utl_limpaCNPJ}
\alias{utl_limpaCNPJ}
\title{Limpeza de CNPJs em um DataFrame}
\usage{
utl_limpaCNPJ(df, nome_coluna)
}
\arguments{
\item{df}{DataFrame que contém a coluna de CNPJs.}

\item{nome_coluna}{O nome da coluna no DataFrame que contém os CNPJs.}
}
\value{
Um vetor de CNPJs limpos.
}
\description{
Esta funcão recebe um DataFrame e um nome de coluna, e retorna uma lista de CNPJs limpos,
com a remoção de caracteres especiais como pontos, barras e hífens.
}
\details{
A máscara do CNPJ deve esta no formato 99.999.999/9999-99.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{https://contabilidados.quarto.pub/}.
Ao acessar, fazer busca pelo nome da função \code{utl_limpaCNPJ}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
\examples{
library(cntdd)

utl_limpaCNPJ(cntdd::dt_cvmB3, cnpj)

}
