% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_limpaCNPJ.R
\name{utl_limpaCNPJ}
\alias{utl_limpaCNPJ}
\title{Limpeza de CNPJs em um DataFrame}
\usage{
utl_limpaCNPJ(df, nome_coluna)
}
\arguments{
\item{df}{DataFrame que contém a coluna de CNPJs.}

\item{nome_coluna}{O nome da coluna no DataFrame que contém os CNPJs.}
}
\value{
Um vetor de CNPJs limpos.
}
\description{
Esta função recebe um DataFrame e um nome de coluna, e retorna uma lista de CNPJs limpos,
com a remoção de caracteres especiais como pontos, barras e hífens.
}
\examples{
df <- data.frame(cnpj = c("00.000.000/0001-00", "11.111.111/1111-11"))
utl_limpaCNPJ(df, "cnpj")
}
