% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_tableStat.R
\name{utl_tableStat}
\alias{utl_tableStat}
\title{Tabela de estatistica por grupos}
\usage{
utl_tableStat(bd, grp1, n_grp1, grp2, n_grp2, value, ...)
}
\arguments{
\item{bd}{Um data.frame}

\item{grp1}{String informando o nome da coluna referente a variavel 1}

\item{n_grp1}{Numero de grupos da variavel 1}

\item{grp2}{String informando o nome da coluna referente a variavel 2}

\item{n_grp2}{Numero de grupos da variavel 2}

\item{value}{Variavel do banco de dados da qual se pretende retirar a estatisca descritiva}

\item{...}{Informar qual estatistica descritiva (mean, sd)}
}
\description{
Calcula a estatistica descritiva de uma variavel numerica baseada em grupos de duas variaveis.
trata-se de uma analise tridimensional
}
\details{
Informacoes adicionais sobre como usar o pacote, orientamos acessar o menu
\code{cntdd} do Blog do Projeto contabiliDados: \url{http://contabilidados.com.br}.
Ao acessar, fazer busca pelo nome da funcao \code{utl_tStarSig}

Contatos pelo email do Projeto contabiliDados:
Email: \href{mailto:contabilidados@ufersa.edu.br}{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \url{https://www.instagram.com/contabilidados} @contabilidados
}
