% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_tableStat.R
\name{utl_tableStat}
\alias{utl_tableStat}
\title{Tabela de média por grupos}
\usage{
utl_tableStat(bd, grp1, n_grp1, grp2, n_grp2, value, ...)
}
\arguments{
\item{bd}{Um data.frame}

\item{col_Value}{String informando o nome da coluna referente à variável númerica sobre a qual se pretende criar os grupos}

\item{col_Grp}{String com o nome da nova coluna a ser criada com os grupos}

\item{n_grp}{Número de grupos, se pretende criar grupos com o mesmo número de observações ou um vetor contendo os percentis para cada grupo. Padrão igual 2.}

\item{lbl_grp}{String com o prefixo posto no início do nome do de cada grupo. Padrão igual "grp"}
}
\description{
Calcula a média de uma variável numérica baseada em grupos de duas variáveis.
trata-se de uma análise tridimensional
}
\examples{

## Grupo 1 com os primeiros 20\% das observações de menor valor, Grupo 2 com as observações de
## 20\% a 80\% e Grupo 3 com os 20\% das observações de maior valor
utl_createGroup(mtcars, "mpg", "grp.mpg", c(0, .2, .8, 1))

## Grupo 1 com metade das observações e Grupo 2 com a metade restante
utl_createGroup(mtcars, "mpg", "grp.mpg", 2)


}
