% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_t_starSig.R
\name{utl_tStarSig}
\alias{utl_tStarSig}
\title{Formata significância}
\usage{
utl_tStarSig(sig, t = NULL, nDigits = 2)
}
\arguments{
\item{sig}{o P-Value}

\item{t}{a Estatística t}

\item{nDigits}{Número de Dígitos}
}
\description{
Formata significância
}
\details{
Essa função formata a estatística t e o nível de significância. Seu resultado
pode ser uma Estatística t com o número de casas decimais informados com o
respectivo p-valor identificado por asteriscos, sendo \code{*}10\%, \verb{**}5\% e \verb{***}1\%.
}
\examples{

Informações adicionais sobre como usar o pacote, orientamos acessar o menu
`cntdd` do Blog do Projeto contabiliDados: \href{http://contabilidados.com.br}{(Acesse Aqui)}. Ao acessar, fazer busca
pelo nome da função `ind_liqSeca`

Contatos pelo email do Projeto contabiliDados:
Email: \email{contabilidados@ufersa.edu.br}
Siga-nos no Instagram: \href{https://www.instagram.com/contabilidados/}{@contabilidados}

}
